Fix elasticsearch plan names to conform to ADR
the es suffix in the name doesn't confirm to the naming ADR[1]. This
updates them to conform.

We had some debate as to whether the version should be expressed as 5.x
or merely 5. We decided to go with 5.x as that makes it more obvious
that this represents a version.

[1]https://team-manual.cloud.service.gov.uk/architecture_decision_records/ADR025-service-plan-naming-conventions/