Run paas-admin custom acceptance tests
Add a task to run the paas-admin integration acceptance tests
in the custom-acceptance-tests job, in parallel with the
existing custom-acceptance-tests.

We run it using the community node container and install the npm
dependencies on the spot.

We did not worry too much about caching of npm modules because
it runs in parallel with the custom acceptance tests, which is
also slow.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>