Add Aiven cost metrics collector and tests.
This commit adds in a new gauge collector in paas-metrics for
Aiven cost of the latest stimated bill, and supporting tests,
into the metrics collector app.

These tests cover cases of selecting the latest bill from the invoices
endpoint, extracting the cost from this json object and the scenario
where an invalid token is sent.

We pass the new variables to access the aiven API when deploying
paas-metrics: AIVEN_API_TOKEN and AIVEN_PROJECT, both already
available in in the pipeline.