Enable Bosh DNS
We have enabled Bosh DNS, as part of this we have taken the oppotunity
to move for using `((environment))` to `((deployment_name))` as this is used
by upstream manifests.

We used to have 000-bosh-rename-deployment.yml which set `name` to
`((environment))`. There's an upstream ops file that does the exact same
thing, but uses `((deployment_name))` instead. Let's use that.

`((deployment_name))` is needed by the use-bosh-dns... thing we're about
to introduce anyway.