Add bosh-dns checks and datadog monitor
We're about to switch from using consul for DNS to using bosh-dns. This
means we'll need to monitor bosh-dns everywhere we used to be monitoring
consul.

This commit adds a `datadog-bosh-dns` check for every machine that had a
`datadog-consul-agent-client` check.

This will alert us in the case that Bosh DNS processes are failing to
either start or keep running.