Aiven service instances cannot change plans
The 2 Elasticsearch plans we're launching with are both the same size,
varied only by ES version (5 vs 6). Also, Aiven doesn't perform a
version upgrade when moving an existing service instance from the
small-ha-5 plan to the small-ha-6 plan.

Taken together, these two facts mean that there's no meaningful plan
change that can be made - so in this commit we mark these plans as
non-updatable.

The above holds, technically, for now, but may not be true in the
future.  We are choosing to delay the work to implement selective
update-ability until we actually introduce differently-sized plans.