Bump log-cache version
We're periodically experiencing responses with `502` on `/v1/meta`
endpoint, and before the issue is raised with the upstream, we'd like to
make sure it doesn't only happen to that specific version of log-cache.

The upgrade also involves other bug fixes and features, including:
- Fix data duplication bug
- Additional metrics for CFAuthProxy

This may solve one of our other stories we have in the backlog.