Increase the collection interval
We're now scraping a lot of metrics every 1 minute. We presume that's
OK, however this includes querying the AWS CloudWatch, which is really
expensive. For the time being in order to save some moneys, we'd like to
run the queries every 10 minutes. A downside is that at the current
state of an application it has to apply these for all metric collectors.

This should be a temporary change, and ideally be reverted back and
sorted in a different way.