Ignore the test causing travis to fail
This is an issue with badssl.com and the fact their certificate expired.
The nature of our monitor relies on it being reported as 0 days
remaining. This would not happen in production systems if we're having
the monitoring in place.

Once the badssl.com has resolved their issue, it should be reverted
back.

Co-authored-by: Daniel Blair <daniel.blair@digital.cabinet-office.gov.uk>