Use upstream opsfile to configure bosh-DNS.
This avoids the need for us to duplicate lots of the config, and makes
it much clearer what we've changed from upstream.

We need to override the aliases for the router and cell VMs because we
deploy these onto a different subnet from the rest of the VMs. The
subnet name and deployment name are included in the DNS names that
bosh-DNS uses meaning that the values in the upstream opsfile aren't
correct here as it assumes all of CF is on the same subnet.