Scale down default dev environment.
This reduces the number of VMs used in the default dev deployment so
that they're not as expensive to run. This does mean that these envs are
no longer HA.

This also introduces a new 'SLIM_DEV_DEPLOYMENT' env var to control this
feature. Setting this to 'false' will restore the full dev environment
for cases where this is needed.