Downgrade log-cache 1.3.0 to avoid memory leak
The log-cache 1.4.1 has some problems freeing memory in the store[1][2]

This has caused some problems in production to us, where log-cache
grew to consume all the memory and swap of the doppler nodes.

This is meant to be fixed in 1.4.2 from comments in slack #log-cache[3]
but it is not ready yet.

The most recent stable version is 1.3.0, it is not as performant but
more stable.

[1] https://www.pivotaltracker.com/story/show/159699403
[2] https://www.pivotaltracker.com/story/show/157700747
[3] https://cloudfoundry.slack.com/archives/CBFB7NP9B/p1534518013000100