Configure test suite name_prefix from config file.
Instead of hard-coding it within the test suite. We currently set this
by writing directly to the config struct within the test suite. This
prevents us from replacing this config struct with an interface type. We
need to do this in order to switch to using the cf-acceptance-tests
config package in our acceptante tests.