Use Config from cf-acceptance-tests instead of cf-test-helpers
The Config in cf-test-helpers uses mixed units - some things have
timeouts in seconds, some in minutes. This is different to the config in
cf-acceptance-tests (changed in [1]) which causes some difficult
situations.

For example, if we try and reuse the same config for the upstream
acceptance tests (which don't use the cf-test-helpers config type) and
our custom acceptance tests we end up with timeouts that are a factor of
60 different.

By using the type from cf-acceptance-tests everywhere we can reuse the
same config and all fields are in seconds.

This change will introduce one difference - the cf-acceptance-tests
config type sets a default TimeoutScale of 2.0, which means that all the
configured timeouts are doubled before being used. As a result, all the
timeouts for the custom-acceptance-tests have effectively been doubled
with this change. We've decided that this is acceptable, as all it does
is bring these timeouts inline with the upstream acceptance tests.

[1] https://github.com/cloudfoundry/cf-acceptance-tests/commit/8637db7