Reduce custom acc tests parallelism in slim dev
We've seen the API get overwhelmed in a slim dev deployment, causing its
healthcheck to fail, which results in the route being unregistered for a
short period. This leads to massive test failures.

Reducing the parallelism here should reduce the load on the api,
avoiding this problem. We don't think the increase in time for these
tests will be a problem because they currently take around half the time
that the upstream acceptance tests take.