Use Eventually when waiting for ES instance
In the custom acceptance tests for the elasticsearch cluster, maybe
it takes a little bit longer to create the test index by the
healthcheck than the time the test takes to check for it,
causing some flakiness.

To fix this we wrap the check of the index around a
Eventually() that would timeout after 5 seconds.