Opsfile to set vm_types to match cloud-config.
The default prometheus manifest specifies a single 'default' VM type
which doesn't exist in our config. We therefore update them to match our
types.

All but the prometheus2 vms have very minimal requirements, so get the
nano VM type. the prometheus2 VM stores the metrics itself (the
database VM is for grafana config alone), so it gets the medium size. It
might need scaling in the future.