Pass a vars store to the prometheus manifest generation script
We have several vars that need to be persisted to avoid recreating VMs
on each deployment. The vars store is bosh's mechanism for doing this.

We're using an envvar to communcate the vars store file with the script
to avoid the "$@"-suffix pattern and ensure the tests reflect the
script's real use.