Terraform to configure ALB etc for prometheus.
We use an ALB with 3 target groups to route requests to the different
ports exposed on the Nginx reverse proxy based on the host header. We've
chosen to do this due to the way that prometheus-boshrelease has laid
out the nginx config - it listens on several ports instead of using
host-based location blocks.