Pin prometheus' stemcell to the same version as CF
We don't want to pin the stemcell to "latest" as this will affect our ability
to rollback stemcells in a hurry. We don't want to incur the overhead of
finding and uploading a separate stemcell from CF, so we re-use the CF ops file
to pin prometheus to the same version.