Ensure persistent disks use the gp2 storage type.
When using `persistent_disk` with a size, you get the AWS default storage type.
This default varies depending on the region you're deploying to[1]. To avoid
this, we use a persistent_disk_type to allow us to ensure we get gp2
everywhere.

We use the 5GB type for the smaller VMs instead of 1GB specified in the
upstream manifest, because the difference in cost is tiny and not worth the
extra definition.

[1] https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html