Improve showing secrets in showenv task
This currently fails noisily if the pipeline hasn't been run, and
therefore the vars-store files don't exist in the S3 bucket. This wasn't
a problem for the existing show-cf-secrets.sh script only because we
hadn't `set -e` etc in it.

This takes the existing script and generalises it to show secrets from
any vars-store file and to fail gracefully if the secrets file doesn't
exsit.