Automate more of the Logit logstash filter creation
There are still manual steps involved but some of the more onerous
manual steps have been removed.

We are now pulling filters directly from upstream releases also.

We had to talk to Logit support in order for the `alter` keyword to be
available for logstash filters. This is available in Logit's logstash
6.x stacks but had to be installed as a plugin manually on our stacks.