Replace Doppler latency monitor
Despite various previous attempts to make this monitor less noisy we
still get spurious alerts. I believe this is because the nature of the
metric we are monitoring: the ELB latency is a very small value (around
9µs) and is subject to high spikes, but not to a level where the latency
is something we would worry about.

This monitor achieves the same result: it gives us an indication of the
health of the Dopplers and whether they are coping with the log and
metric volumes.

The monitor is a change alert because the metric is a count (and will
therefore keep increasing for the lifecycle of the VM). It checks the
maximum count in the last 10 minutes and compares it to the value an
hour ago[1].

[1] https://docs.datadoghq.com/api/?lang=python#create-a-monitor