Fix install buildpack ref check test: cfdep v3.6.0
In one of the releases between v2.6.0 and v3.6.0 there has been a change
in what packages (and their names) that are included in certain jobs.

Now instead of the release 'staticfile-buildpack' containing a job
'staticfile_buildpack' with the package 'staticfile-buildpack'

There are multiple packages:
- 'staticfile-buildpack-cflinuxfs2'
- 'staticfile-buildpack-cflinuxfs3'

So we have modified the test to account for this.

It would be too expensive when running the tests to pull everything down
and check that the packages exist, and we have other testing
infrastructure that will check this rarely modified convention.

(Everything will blow up if these packages do not exist)