delete no-op opsfiles for cf-deployment v3.6.0
In cf-deployment v3.6.0 the versions of loggregator and cflinuxfs2 are
now the same as the versions that we had in our opsfiles, therefore the
opsfiles are redundant. We can get rid of the opsfiles and have the same
functionality.

If we need to change the versions (e.g. CVE mitigations) we can pull
these files out of git or re-add them.