Mark HA elasticsearch plans as non-free.
We don't want users on a trial quota to be able to use the more
expensive plans, so this marks them as non-free. When this is not
specified, CloudFoundry defaults to marking plans as free[1]. Specify it
for all plans to make it more obvious what the behaviour will be.

[1] https://github.com/openservicebrokerapi/servicebroker/blob/master/spec.md#plan-object