Trigger prod pipelines from staging in London
We are removing our staging environment in Ireland. This was the
environment that ultimately triggered the deployment of production in
Ireland.

This change will effectively detach staging in Ireland from the
pipelines which should allow us to tear down this environment without
stopping deploys.

We will look to tidy up the naming inconsistencies later.