Make staging London tag with `prod-{version}`
We used to deploy to production London and production Ireland from a
staging environment in the respective region. We now only have a single
staging environment in London. We're changing staging to use a tag that
does not include the region name. This is because seeing production
Ireland consuming tags for "prod-lon" reads very strangely.