Always use Ireland for SES SMTP server.
We were interpolating the AWS region into the hostname being used here,
but SES is only available in 3 regions globally[1] - London isn't one of
them.  email-smtp.eu-west-2.amazonaws.com does not resolve.

This causes the password reset flow to fail with a 500 error because UAA
is unable to mail the reset link to the user.

We address this by hardcoding eu-west-1 (Ireland) as that's the closest
available region for all our deployments

[1] https://docs.aws.amazon.com/ses/latest/DeveloperGuide/regions.html