Enable elsticsearch service for everyone.
We want to allow all users to be able to use this service. This adds the
call to the pipeline to enable it for everyone.

I've also updated the tests to not enable this service for the test org
so that we can verify it's on for all orgs by default.