Be more explicit in defining elasticsearch test plan
The way this test selected the plan to use was non-obvious to me, and at
least one other team member. I've therefore refactored this to be
explicit about the plan that's being used in the test.

I've also changed it to use an elasticsearch 6.x plan to test as we
recommend that tenants use this version unless they have a good reason
to use 5.x.