Fix cdn cert monitor to not alert for individual invalid certs
This was added in c3066ec with the intent of only alerting if a
significant number of certs were invalid, not when a single one was, so
that a cert that's in the process of being provisioned doesn't trigger
this monitor. It was intending to take the average of the values (which
are reported as 1 or 0 for each hostname), and alert if that fell below
a threshold.

However, because this was applying a `min` before doing the
`rollup(avg)`, it was taking the average of the minimum value, and hence
would always evaulate to 0 if any certs were invalid.

I've changed this to use the `avg` instead, and get rid of the rollup.
This results in a floating-point value representing the percentage of
certs that are valid. I've also changed the thresholds because a single
invalid cert in prod currently drops the value from 1 to 0.98, so these
thresholds seemed overly lenient