Remove test for install_buildpacks packages.
This test was added in 5c6abfe in response to us missing a change to the
java buildpack name. The test was never fully correct as it relied on
the buildpack job names matching the package names, which was never
guaranteed.

With the introduction of the cflinuxfs3 stack, most of the buildpacks
have now been updated to include 2 packages - one for each of the 2
stacks (cflinuxfs2 and cflinuxfs3) with the stack name added as a suffix
to the package name.

It's not possible to implement this test correctly without either
downloading and extracting the buildpack releases, or replying on
scraping bosh.io for information about the packages they contain.
Additionally, now that we're directly consuming upstream's manifests,
the risk of missing future changes is greatly reduced. That coupled with
the procedure of testing CF upgrades on a clean build from scratch
should mean that this test is now far less necessary than it was.