Deploy Metric Exporter
We want to have the same metrics provided by the paas-metrics-exporter
available via log-cache, and therefore via the adapter to Promethii. To
achieve this the paas-metric-exporter was adapted to be able to send to
Loggregator. This requires writing to a local Metron Agent, and
therefore is deployed as a Bosh release.

It has been colocated with the log-api instance group for cost saving
purposes.

It has a metric whitelist configured because it seems
Loggregator/Log-cache currently supports the container metrics. Note:
this means we are not going to be producing the `crash` count metric.

Regarding the OAuth authorities of the UAA client: this client has to be
capable of reading a list of apps from the CF API and reading from the
firehose. I initially just used `cloud_controller.global_auditor`, which
worked for getting the apps, but gave an error when trying to connect to
dopplers. This was confusing, because it also had `doppler.firehose`,
which I thought was sufficient for reading from Dopplers. I didn't have
the time to launch an investigation into why this is. Instead, I used
`cloud_controller.admin_read_only`, which Just Works.