Use autopilot for healthcheck app deployment
We've seen several occasions where the blue-green plugin will fail to
detect that the app already exists when deploying a dev environment. The
deploy then fails at the step when it attempts to rename the -new app
because the target appname is already taken.

I've not got to the bottom of why this happens, but switching to
autopilot has removed the issue for me. This plugin also has the
advantage that it purely relies on the manifest.yml for details of the
app, and doesn't attempt to preserve any details from the existing
deployment.