Modernise the plugin
TL;DR We're changing the script to meet latest "cool" standards accepted
by GDS.

We're using `const` instead of `var`, to determine the elements never
being changed.

Converting the `function() {}` to shorter new version `() => {}`, just
so my linter won't go crazy.

We're renaming the `element` variable name, which is being overwritten a
lot, to something more meaningful and unique. Prefixing the variable
name with `$` sign, in order to defrienciate DOM elements, from "real"
JS variables.

The `location` isn't really a thing. We should be using
`window.location` instead.

We're using the built in templating "engine", in order to prevent word
concatonation - good practice if anything.

Concourse uses Elm language for it's JavaScript implementation. It is
very picky, when you start deleting things it has registered before.
Instead of deleting these elements, we're going to hide them, and create
a new element that will print our hostname on the bottom of the screen.

The [`font`][font] tag isn't really a thing anymore. We're replacing it
with a `div` instead and adding some inline styles.

Also, converting to 2 space indenting.

[font]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/font