Hide an extra element
Now that we have upgraded concourse, some of the frontend has changed.
This means our plugin is out of date and should be updated if we'd like
to keep on using it on our dashboard TV the way we did.

In the new concourse version, upstream has decided to separate the
grouping bar into it's own place. This is a simple change really and
requires some minimal effort to hide new additional element.