Scope RDS disk utilisation monitor to region
Currently if the RDS ultilisation monitor is triggered it will show for
both "prod" and "prod-lon". This is because it was only scoped to the
aws-account and we now have multiple things in each account.

To make this more intuative, this change scopes the monitor to the same
region as the deploy_env. Since we do not expect to have multiple
deploy_envs within the same region this should result in alerts being
triggered for their respective deploy_env.