Add job for rotating CC DB encryption key
We want to be able to update the CC DB with the name of the key which
has been used to encrypt the values in it.

In the database's current state it is missing records for which key was
used to encrypt the values. This job is being added only temporarily to
fix the missing data. See [1,2] for more details.

Eventually, when we work out how we will do key rotations, the task in
this job should probably be run in the rotate-cloudfoundry-credentials
job. For now it is a standalone job to allow it to be run without
forcing you to do a full rotation.

[1] https://github.com/cloudfoundry/cf-deployment/pull/561
[2] https://github.com/cloudfoundry/cf-deployment/releases/tag/v2.9.0