Enabled instance identity
We want to enable 'instance identity', which means allowing the rep to
be able to provide application container with their own certificate.
This does _not_ actually enable HTTPS traffic between GoRouter and the
apps. For that, you have to enable the Envoy proxy. Given the Envoy
proxy is still experimental we have decided not to enable it yet.