Revert "Merge pull request #1559 from alphagov/deploy-metric-exporter"
This reverts the changes introduced in the PR #1559[1].

This was causing performance problems in the loggregator stack.
The reason is that the current implementation of
paas-metrics-exporter [2] is designed to monitor for any new app
accessible for the given credentials, and opening a new WebSocket
stream for each of them.

This works fine for tenants monitoring their own apps, but for
us with admin credentials means creating streams for ALL the
applications, thousands of them.

The paas-metric-exporter was eventually running out of file
descriptors and failing with errors like this:

       fetching events failed: Error dialing trafficcontroller server: dial tcp 34.249.189.44:443: socket: too many open files.
       Please ask your Cloud Foundry Operator to check the platform configuration (trafficcontroller is wss://doppler.cloud.service.gov.uk:443).

Possible, as it was failing, it tried to reconnect later on.

This was hammering the loggregator stack, by reconnecting the
traffic controller, and this one proxying to the dopplers.

Dopplers suffered an increase in CPU usage of 35-40% and thousands
of logs reporting connections/disconnections both in traffic
controller and in the dopplers.

We disable this for the time being, meanwhile we decide a follow-up,
like fork the existing paas-metric-exporter or adapt it.

[1] https://github.com/alphagov/paas-cf/pull/1559
[2] https://github.com/alphagov/paas-metric-exporter