Update cflinuxfs2 to 1.242.0 to fix git CVE
Update to latest version of cflinuxfs2 to cover CVE USN-3791-1[1]

Git vulnerability

It was discovered that git did not properly validate git submodule urls
or paths. A remote attacker could possibly use this to craft a git
repository that causes arbitrary code execution when recursive
operations are used.

CVEs contained in this USN include: CVE-2018-17456

[1] https://www.cloudfoundry.org/blog/usn-3791-1/