Add bosh-dns to manifest addons
We want bosh-dns as an in-manifest addon, rather than in the runtime
config, because it makes monitoring of certificates and certificate
rotation much easier. Also, there is no actual requirement for bosh-dns
to be in runtime config.

We also separate the bosh-dns and bosh-dns-alias ops files. This is
because they are in separate Bosh releases, so it makes it clearer where
releases get their configuration.

Note: the bosh-dns-alias ops file is required because the cells and
router run on different networks to the rest of the CF stack.