Remove service credentials credhub
A credhub job was introduced by upstream[1] to store credentials for
service brokers, instead of having them stored in the CCDB.

We want to delay adding credhub to our stack, especially at the CF
level, until we have played the spike in [2] and understand it better.

We do not use any service brokers which require/allow credhub
integration, so this is safe until further cf-deployment integration
forces us to re-evaluate - or until we choose to do so based on any
spikes.

Initial reading[3] has revealed that in order to use credhub-backed
service credentials, the deployed applications themselves need to
interact with credhub to retrieve the credentials. This is unlikely to
fly with our users without significant help from us, or perhaps
auto-interpolation through the lifecycle management mechanism.

[1] https://github.com/cloudfoundry/cf-deployment/releases/tag/v4.0.0
[2] https://www.pivotaltracker.com/story/show/158978139
[3] https://github.com/cloudfoundry-incubator/credhub/blob/master/docs/secure-service-credentials.md