Add logic to pause the kick-off after kick-off
We want to not trigger the deploy pipeline automatically for every
environment unless the developer opts in the day before.

To do so, we simply add an additional job that would pause the
timer that triggers the kick-off after the timer is triggered.

The developer must unpause the resource manually to get the deployment
the next day.

We also introduce the variable `makefile_env_target` when calling
`make` from the `deployment-kick-off` pipeline, to keep consistency
with other pipelines and avoid forgetting any variable.