Fix install_buildpacks for older binary_buildpack
We're not upgrading this at the same time as upgrading the rest of CF.
The newer version has introduced support for cflinuxfs3, and as a result
has renamed the packates contained within. cf-deployment has been
updated to reflect this change. As we're not upgrading it, we need to
set the value back.