Script the creation of orgs, not users
Since we've got paazmin now we should probably use that for inviting
users and we can avoid a bunch of complexity in this script. We don't
have the ability to add new orgs in paazmin yet though, so we still need
a script for that.

The script now just creates an empty org / space and prompts the user to
go to paazmin and invite the new users instead of scripting it all.

The script will be renamed in a subsequent commit so it's easier to
review this.