Use YAML anchors to specify pipeline containers.
We can avoid a lot of repetition in the pipelines by defining the
details of the image_resources we use several times once at the top of
the file, and then referencing them using YAML anchors. The top-level
`meta` key in the YAML will be ignored by Concourse when we do
set-pipeline, and is only there to hold anchor definitions.