Fix the pipeline kick-off pausing in London
When the pipeline pausing was introduced in d3fee14, the
`AWS_DEFAULT_REGION` param was removed from these 2 tasks. This causes
them to fail when not in eu-west-1 (Ireland) as the s3 API returns a 400
error when trying to access the bucket in the wrong region (called from
the `environment.sh` script to get the concourse password[1]). This is
because the Makefile defaults this to eu-west-1 if it's not set[2].

[1]https://github.com/alphagov/paas-cf/blob/bb5bf1e/concourse/scripts/environment.sh#L33
[2]https://github.com/alphagov/paas-cf/blob/bb5bf1e/Makefile#L95