Enable monitor bosh in prometheus deployment
We add the bosh_exporter with alerts and dashboards.

For the time being we configure the bosh_exporter to use the same
unique admin user from bosh, reading directly from the bosh-secrets.yml.

We pass these credentials into a YAML vars-file in the prometheus
generate-manifest.sh.

In the future we should use a readonly user as described in the
upstream opsfile[1], but this requires UAA in the bosh deployment.

[1] https://github.com/bosh-prometheus/prometheus-boshrelease/blob/master/manifests/operators/bosh/add-bosh-exporter-uaa-clients.yml