Reorg rds-broker operation files
Break down the rds-broker configuration in different files and
specific operations:

 - One file to setup the rds-broker instance_group and related jobs
 - One to setup the postgres service and plans
 - One to setup the mysql service and plans
 - Add each plan in separated bosh operations

The aim is to increase readbility and make it easier to understand
and change, without having too many levels of intentation on a
YAML file.

The drawback of splitting is that we had to duplicate the anchors,
as YAML anchors are applied per .yml file.

We also change the test to use a manifest without vars-store, which
speeds up running the specific RDS spec file.