Custom test for wildcard host for shared domains
We want to be sure that we have a wildcard domain configured for
any of the non-internal shared domains in the platform, and that
this wildcard domain is set in the "govuk-paas" org.

The reason is that we do not want any tenant to configure a wildcard
host on the default shared domains. Instead, we want to decide
what we do with them.

Internal domains do not allow wildcard domains.