Add task to configure wildcard routes
We want to pre-assign a host '*' route for the main shared domains
(non-internal) into our govuk-paas org, so that no other tenant would
be able to register the wildcard domain.

This would prevent a potential issue if someone registers the wildcard
for any other application, potentially hijacking the traffic of apps
being unavailable, or any route not mapped yet.

For the time being the route remains unmapped, in the future we
might want to assign this to a custom application with some message
and custom HTTP status code.