Add CF and Firehose exporters to Prometheus
We want to start putting Cloudfoundry metrics into Prometheus. We try to
stick to the upstream ops files as much as possible to make it easier to
do upgrades.

The deployment relies on secrets generated when deploying Cloudfoundry
(UAA client secrets). In hindsight, we probably should have used a
single vars store for both CF and Prometheus.

We hardcode `traffic_controller_external_port` because it is hardcoded
everywhere else.