Remove UAA client unit test
So far all this test has done for us is break every time we add a new
UAA client. I am not sure what this test is protecting against. It is
not enforcing any kind of policy on the structure of our manifest.
Removing to save the maintenance burden of keeping the list up-to-date.