Set network and VM type for Prometheus firehose
These operations were not added to the existing files[1] because those
are applied immediately after the base manifest, at which time the
firehose instance group does not exist. This is an ordering problem. I
thought about moving the files around, but it makes sense to keep
separate ops files anyway, because it makes features easier to
add/remove simply by adding/removing the necessary files.

[1] 100-update-networks.yml and 100-update-vm-types.yml