Check that platform tests compile in test
The staging pipeline failed because of a missing colon in the platform
tests - this sort of thing should be caught at unit test time. Because
we can't run the platform tests, we weren't even trying to compile them.

Using `go test -run ^$` makes go compile the test code, but doesn't run
any tests (because they don't match the empty string pattern).

Also fixing the missing colon that made the tests fail and an unrelated
error about `%s` that occurs with my version of the comiler.