Add a script to reset users' passwords
In https://github.com/alphagov/paas-cf/pull/1584 we changed the
create-users.sh script to just create orgs, not users.

It turns out that script was also the canonical way of resetting users'
passwords -
https://team-manual.cloud.service.gov.uk/guides/common_support_tasks/#changing-a-user-39-s-password

I don't think it makes sense to use the same script for both tasks, so
here's another script that does more-or-less the same thing, but relies
on paas-admin to do the invite sending.

See
https://github.com/alphagov/paas-cf/blob/3579e0aa9f2c171b95be629a00a3c0b9c10fd8ee/scripts/create-user.sh
for the old implementation.