Remove deployment-ssh-keypair from pipeline
We were generating this keypair in the cloudfoundry pipeline if it
didn't already exist, but it's not used in this pipeline at all. It's
also generated in the create-bosh-concourse pipeline[1], and used there,
so it will always exist by the time this pipeline runs meaning that this
job is always a no-op.

[1]https://github.com/alphagov/paas-bootstrap/blob/900d55b/concourse/pipelines/create-bosh-concourse.yml#L623