Simplify cf generate-manifest.sh with slim-dev-env
Now consul is not there in cf-deployment and we do not have a opsfile
to remove consul, we can add scalet-to-one-az.yml at the same variable
than other opsfiles.

It must be first before our custom opsfile.

We also speed up the corresponding tests by no using vars-store
in the tested manifest, so it does not generate certs.

[1] https://github.com/alphagov/paas-cf/commit/e24a3177#diff-31c9fe7d5b430915c2e9db0a24e645c8