Use Postgres 10 in custom acceptance tests
We have unit and integration tests for Postgres 10 in the RDS broker,
but there are no acceptance tests for it. We have high confidence in the
broker's support for Postgres 9 already, so we should switch to Postgres
10 by default from now on.

I did not add a Postgres 10 test on top of the existing one because it
adds more complexity to the test, and may increase the run time. The
integration tests already cover most of the broker functionality anyway,
these custom tests are more about the end-to-end connectivity and
usability.