Change max_in_flight to 30% and canaries to 1
We want to speed up our production deployment by paralelising the
updates of diego-cells.

For that we change both the `max_in_flight` and `canary` values for
bosh deployments. We have based the values by calculating the amount of
memory that is used factoring overcommiting and as a result how many
cells we can safely roll at the same time.

We found out the right number based on the following calculations
of the expected consumed memory:

    102720Mib total org quotas
    1171627Mib App usage
    30662MiB cell size

    Cell Size * OverCommitment factor = Cell Size with overcommitment

    30662 * 1.66 = 50898.92
    ---
    50898.92Mib Cell Size with overcommitment

    App usage / cell size with overcommitment = cells needed

    1171627 / 50898.92 = 23.01

    23.01 cells are required with overcommitment to satisfy our needs.

The group of VMs most affected by this are diego-cell in Ireland. At the
moment we have 39 instances, 13 instances per AZ.

Bosh deploys the VMs AZ by AZ, applying `max_in_flight` for each AZ. To
calculate the speedup, we assume the same time to update each VM, so the
speed up would be `num_cells / roundup(num_cells/threads)`

For different values of % of  `max_in_flight` (computed based on [1])

| MiF % | total | nodes/AZ | MiF/AZ | speed up |
|-------|-------|----------|--------|----------|
|   10% |    39 |       13 |      1 |       0x |
|   20% |    39 |       13 |      2 |      ~2x |
|   30% |    39 |       13 |      3 |      ~2x |
|   40% |    39 |       13 |      5 |      ~4x |
|   50% |    39 |       13 |      6 |      ~4x |
|   60% |    39 |       13 |      7 |      ~6x |
|   70% |    39 |       13 |      9 |      ~6x |
|   80% |    39 |       13 |     10 |      ~6x |
|   90% |    39 |       13 |     11 |      ~6x |
|  100% |    39 |       13 |     11 |     ~13x |

Based upon the above data a `max_in_flight` value of 30% should be used
to gain a doubling of deployment speed when rolling cells.

As we will be rolling more than one cell at a time we are using the
`canary` option with a value of 1. This ensures that the new cell is
operating properly before rolling larger numbers.

The rest of instance groups will not be affected by this change as we
run in general 3 instances of each type. For `doppler`, we run
6 instances at the moment, but the computed MiF per AZ will remain 1.
Having `canaries` of 1 with MiF of 1 is equivalent to 0 canaries.

[1](https://github.com/cloudfoundry/bosh/blob/master/src/bosh-director/lib/bosh/director/deployment_plan/update_config.rb#L128))