Scale API instances in production to 4
... and make sure they span three AZs.

We had an incident recently where enabling the prometheus cf_exporter
caused some API requests to time out. We think this wouldn't have been a
problem if the API instances had been scaled up.

We're going with 4 instances (rather than 3) so that there's enough
of a difference for us to notice any changes in our metrics.

We're not planning on re-enabling the cf_exporter as part of this story.