UAA: allow emails with more than 4 domain parts
The default allowlist for UAA's internal IDP only allows users with 4
parts in their email domain[1]. This means that UAA returns an error
when we (or an Org Manager) attempts to invide a user with 5 or more
components in their email domain.

It doesn't appear to be possible to configure this via UAA's config file
(as per a conversation on CF slack[2]), and therefore via the manifest,
so instead we have to add a post-deploy job to update this via UAA's
API. This extends the list to allow up to 6 components in the domain.
This should be enough for now, and it's reasonably simple to extend
further if needed.

In testing we have found that this setting works, but gets reverted each
time a UAA instance starts, which is less than ideal. We've tested that
this only affects inviting users, and that they can still login when
this gets reverted, so we consider this is an acceptable workaround for
the time being until a more permanent fix can be implemented in UAA
itself.

[1]https://github.com/cloudfoundry/uaa/blob/4.24.0/server/src/main/java/org/cloudfoundry/identity/uaa/util/DomainFilter.java#L103
[2]https://cloudfoundry.slack.com/archives/C03FXANBV/p1542628288065500